2024 day 6
HOLY CRAP I FINALLY DID IT AFTER 5 HOURS!!!!
Couple of stupid mistakes in part 2:
- tried to actually solve it with an algorithm, turns out there are too
many edge cases for me to cover.... so bruteforcing!
- my retarded ass forgot to implement a real fool-proof function to
detect if the guard is out of bounds.. an hour wasted to debug this
stupid mistake.
- tried to multithread the solution after it took too long to finish
even the first branch... another wasted hour, turns out my stupid
implementation only detected an infinite loop if the guard returned to
the INITIAL SQUARE. shoutout to Sebastial lague cuz i somehow noticed
this while watching his latest video.
- after carefully analyzing the example solution, one hour wasted, i
noticed that we can't branch if we've visited this square before.
thankfully this was the last bug.