2024 day 11
only managed to solve the first part, the second part is actually fully
working, though in its current state, it's going to take an eternity to
even come close to 75 passes, even with multithreading. the only
possiblility is to hash the "pebbles", but i do not have neither the
time nor the patience to do any of that, so, for now, i'm fine with
"theoritically working".