finished day 11 part 2
was able to reduce the depth from 75 to 39, since it only takes like a
minute to get all the pebbles using the first part's solution. then
we're discarding the dublicate pebbles, and then hashing the pebble
count after applying the rest of the 75 passes (36) on each of the
pebbles. these hashes are then used to count the number of items that we
would've gotten if we were to use the algorithm with full depth. this
possibly saves freaking eons of computation time and terabytes of
dedicated ram, and grants me one extra star this year :).