2024 day 16 part 1
part 2 is hard as fuck i couldn't solve it. i have a semi-functional
algorithm to get all possible paths, which is to first get the path with
the least cost, and then backtrace through all the nodes, and if there's
an intersection between two paths, we spawn another dijkstra search to
get the smallest distance from that point, and if that distance plus the
cost of the backtraced path is equal to the best path cost, we can
safely add that to the list of shortest paths. one small problem, while
dijkstra is finding the shortest path, it may cut off the nodes in front
of the new path and just get one of the paths we've already looked into,
which means a lot of missed nodes >:(